backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
passlib[bcrypt]
python-jose[cryptography]
httpx
python-dotenv
pydantic-settings
python-multipart
python-socketio
python-engineio
boto3
